Fix shared library builds for Windows
If extensions are built as shared libraries, `PHP_EXPORTS` is not
defined for the extension on Windows, so the `PHPAPI`s are not
exported.  We could fix that by adding a respective define to the
extension build, but that would not allow other extensions to import
the symbols later.  Therefore, we have to define our own export marker.